+++ src/Autocomplete/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 -   Fix issue where `max_results` was not passed as a Stimulus value (#538)
+-   Add all possible stylesheets for tom-select to the autoimport to choose from
 
 ## 2.5.0